requirements.txt
certifi==2019.6.16
cffi==1.12.3
cloudpickle==1.0.0
cycler==0.10.0
Cython==0.29.7
cytoolz==0.9.0.1
dask==1.2.2
decorator==4.4.0
imageio==2.5.0
kiwisolver==1.1.0
matplotlib==3.0.3
mkl-fft==1.0.10
mkl-random==1.0.2
networkx==2.3
numpy==1.16.2
olefile==0.46
pandas==0.24.1
Pillow==6.0.0
protobuf==3.7.0
pycparser==2.19
pydensecrf==1.0rc2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.3
scikit-image==0.15.0
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
tensorboardX==1.6
toolz==0.9.0
torch==1.1.0
torchvision==0.3.0
tornado==6.0.2
tqdm==4.31.1
